be/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
ruff==0.1.14
autopep8==2.0.4
requests==2.31.0
gunicorn==21.2.0
gevent==24.2.1
stripe==12.0.1
celery==5.5.2
django-celery-results==2.6.0
web3==6.9.0
cryptography==36.0.2
google-cloud-storage==3.1.0
